docs: ☎️ Add a support page to the website (#102)
The app was rejected from App Review because there wasn't a dedicated
support flow on the website. This change moves the different sections
back into separate pages to make them clearer when linking to them and
adds a new dedicated support page with links to view GitHub Issues,
raise a new issue, and email.